requirements.txt
amqp==5.0.6
amqplib==1.0.2
anyjson==0.2.5
asgiref==3.4.1
billiard==3.6.4.0
carrot==0.10.7
celery==5.2.0
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-timezone-field==4.2.1
djangorestframework==3.11.0
greenlet==1.1.2
kombu==5.2.1
prompt-toolkit==3.0.22
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
